requirements.txt
# FastAPI WhatsApp Restaurant Bot - Requirements
# Production-ready dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Twilio WhatsApp Integration
twilio>=8.10.0

# Environment & Configuration
python-dotenv>=1.0.0

# Data Validation
pydantic>=2.5.0

# HTTP Client (for webhooks)
httpx>=0.26.0

# Async Support
anyio>=4.2.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Development Dependencies (optional)
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# black>=24.1.0
# isort>=5.13.0
# mypy>=1.8.0
